+++ api-builder-plugin-fn-elasticsearch/src/index.js
@@ -1,6 +1,8 @@
 const path = require('path');
 const { SDK } = require('@axway/api-builder-sdk');
-const actions = require('./actions/search');
+const { search } = require('./actions/search');
+const { getTemplate, putTemplate } = require('./actions/indexTemplate');
+const { getMapping, putMapping } = require('./actions/indexMapping');
 
 /**
  * Resolves the API Builder plugin.
@@ -14,7 +16,8 @@ const actions = require('./actions/search');
  */
 async function getPlugin(pluginConfig, options) {
 	const sdk = new SDK({ pluginConfig });
-	sdk.load(path.resolve(__dirname, 'flow-nodes.yml'), actions, pluginConfig);
+	sdk.load(path.resolve(__dirname, 'flow-nodes.yml'), {search, getTemplate, putTemplate, getMapping, putMapping}, pluginConfig);
+	//sdk.load(path.resolve(__dirname, 'flow-nodes.yml'), , pluginConfig);
 	const plugin = sdk.getPlugin();
 	//plugin.flownodes['elasticsearch'].methods.search.action = actions.search.bind({pluginConfig});
 	return plugin;