+++ api-builder-plugin-fn-elasticsearch/test/basic/setupElasticsearchMock.js
@@ -1,18 +1,24 @@
 const simple = require('simple-mock');
 const fs = require('fs');
 
-function setupElasticsearchMock(client, fn, responeFilename, shouldError) {
-    simple.mock(client, 'isMocked', true);
-    if(!shouldError) {
-        const mockedFn = simple.mock(client, fn).callbackWith(null, 
-            JSON.parse(fs.readFileSync(responeFilename)));
+function setupElasticsearchMock(client, methodName, responeFilename, shouldError) {
+    var mockedFn = simple.spy(function (params, options, callback) {
+        const resonse = JSON.parse(fs.readFileSync(responeFilename), null);
+        if(shouldError) {
+            // Return the given response as an error
+            callback(resonse, null);
+        } else {
+            // Otherwise return the response as expected
+            callback(null, resonse);
+        }
+    });
+
+    // Use the extend functionality of the ES-Client to register the mocked method
+    client.extend(methodName, { force: true }, ({ makeRequest }) => {
         return mockedFn;
-    } else {
-        const mockedFn = simple.mock(client, fn).callbackWith(
-            JSON.parse(fs.readFileSync(responeFilename)), 
-        null);
-        return mockedFn;
-    }
+    });
+    // Return the mocked function to perform assertions
+    return mockedFn;
 }
 
 module.exports = {